+++ imagedl/modules/sources/__init__.py
@@ -5,13 +5,14 @@ from .i360 import I360ImageClient
 from .baidu import BaiduImageClient
 from .google import GoogleImageClient
 from ..utils import BaseModuleBuilder
+from .pixabay import PixabayImageClient
 
 
 '''ImageClientBuilder'''
 class ImageClientBuilder(BaseModuleBuilder):
     REGISTERED_MODULES = {
         'BingImageClient': BingImageClient, 'BaiduImageClient': BaiduImageClient, 'GoogleImageClient': GoogleImageClient, 
-        'I360ImageClient': I360ImageClient,
+        'I360ImageClient': I360ImageClient, 'PixabayImageClient': PixabayImageClient,
     }