+++ tests/test_image_clients.py
@@ -10,7 +10,7 @@ from imagedl.modules.sources.bing import BingImageClient
 from imagedl.modules.sources.i360 import I360ImageClient
 from imagedl.modules.sources.baidu import BaiduImageClient
 from imagedl.modules.sources.google import GoogleImageClient
-
+from imagedl.modules.sources.pixabay import PixabayImageClient
 
 # bing tests
 client = BingImageClient()
@@ -26,5 +26,9 @@ image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)
 # google tests
 client = GoogleImageClient()
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
+client.download(image_infos, num_threadings=1)
+# pixabay tests
+client = PixabayImageClient()
 image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)