+++ docs/Changelog.md
@@ -1,6 +1,6 @@
 # Release Log
 
-- 2025-10-19: Released pyimagedl v0.1.4 — Add a deduplication feature and support 360 image search.
+- 2025-10-20: Released pyimagedl v0.1.4 — Add a deduplication feature and support 360 image search.
 
 - 2025-10-19: Released pyimagedl v0.1.3 — code cleanup, deprecated/invalid functions removed, new functions added.